change package structure to separate src & tests
